backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic-settings
python-dotenv
email-validator
passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt==3.2.2
